Dont err on eval, butr FAIL
This commit modifies the CEL evaluator returns to not err
on an evaluation error. If a policy has bad CEL syntax, we
error. But if it has good syntax but eval fails, we simply
fail the policy. This makes it easier to deal with missing
fields as a policy will simply not pass if data is missing.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>